[VPJ] Refactor some functions and moved GenericAvroRecordReader.java to a diff package
